Return `DeclRef` from `TypeBinding::type_check`. (#4308)
## Description

The main goal of this PR is to refactor
`TypeBinding::type_check_with_ident` into a new generic trait
`TypeCheckTypeBinding<T>` with a function `type_check` that returns
`DeclRef<DeclId<T>>`. The fact that this is done with a generic trait is
an implementation detail and is a workaround to create easy generic
returns in Rust. The important bit is that this function now returns a
`DeclRef<_>` instead of a `TyDeclaration`.

In doing this I discovered a bug with the implementation of
`CollectTypesMetadata` on `TypeId`. The implementation did not account
for any generics that might appear in _nested_ positions of types (like
inside the types of enums, structs, etc). To fix this, this PR
introduces a new abstract method `extract_any` on `TypeInfo` with method
signature: `fn extract_any<F>(&self, .., filter_fn: &F) ->
HashSet<TypeId> where F: Fn(&TypeInfo) -> bool`. The `extract_any`
extracts any bugs that match the criteria for `CollectTypesMetadata`.

This PR also updates some of the previous methods on `TypeInfo` to use
the new `extract_any` method.

- Easy generic returns in Rust:
https://blog.jcoglan.com/2019/04/22/generic-returns-in-rust/

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [x] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.