Update `DeclEngine` API to use more abstract methods. (#4315)
## Description

This PR updates the `DeclEngine` API to use more abstract
implementations of `get`, `insert`, and `replace`. The goal of this PR
is to move the `get` method into a separate trait so that there can be
individual implementations for `DeclId` and `DeclRef`. In this PR there
is no need to introduce different behavior between the two
implementations, but there will be in the future.

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [x] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.