feat: relaxed salt implementation for forc deploy
This allows specifying multiple salts for when deploying workspaces.

This is done through a parse and validation step, where we match
given salts and throw any errors if:

1) Salts were given in an invalid format, or
2) Conflicting salts were given, such as when a salt exists under
contract dependencies and still given in the CLI input.

Tests are also included in this change in the form of plain manifests
under named folders.

remove dangling comment

make deploy_pkg take a ref to ContractSaltMap instead

remove redundant nested standalone_contract