Allow any `Ident` when parsing `asm` block instruction opcodes (#4431)
## Description
The idea is to not check for Sway keywords when parsing the opcodes in
`asm` blocks. This is important because some Sway keywords, such as
`mod`, are also FuelVM opcodes. Note that the validity of the opcodes
are later checked anyways so we're not losing any useful checks here.

## Checklist

- [ ] I have linked to any relevant issues.
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [ ] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.