chore: deprecate built_pkgs_with_manifest()
@kayagokalp rightfully mentioned that the package descriptor is already
contained within the BuiltPackage struct, which means we do not need
this representation anymore:

This is a nice abstractions but it made me think a little and I think
returning Vec of (PackageManifestFile, Arc<BuiltPackage>) might be unnecessary
at the first place from built_pkg_with_manifest function.

Every built package contains a package descriptor which contains the related package manifest
so returning a tuple is kind of a data duplication.