fix: insert a NewlineSequence instead of Vec<NewlineSequence>.
This contains quite a big change for how we insert newlines into the formatted code - currently we would get a Vec<NewlineSequence> to insert into a span, and this PR changes this to just a NewlineSequence.

The bug was a result of collecting this Vec stated above when checking between the last unformatted leaf span and the end of code leaf span that was manually inserted. We then end up inserting it all in one place - resulting in a bunch of newlines rather than just 1.

The reason for changing this to just a single NewlineSequence is because IMO there isn't really a situation where we might have multiple newline sequences between 2 leaf spans such that we want to insert a Vec<NewlineSequence>, but it's causing problems when it comes to the situation described above.

To get a look at how this situation looks like - the test case added shows an example of this.