ci: extract out verify_tags step from publish job into its own job
Further explanation is on #4495, but the gist of it is that sometimes
the verify tags step may fail but we do not get this feedback until
all other jobs are already run since the check has all the other jobs
as a pre-requisite.

Extracting this step out would give us immediate feedback instead of
requiring us to wait 20-30 minutes. We may also add more pre-publish
checks here.