Fix `--finalized-asm` output and docs
This change does three things.

First it repairs the output of the `--finalize-asm` build option so it
actually outputs readable asm.

Second it updates the documentation on build profiles and in all
occurences of `--print-finalize-asm` to use the new flag, use the
correct names of build profile options and adds documentation for the
new flags. Fixes #4481

Third, it lets the deserialization of the flags to use default values
when the flags for a certain build profile are not all defined, so users
can define a new build profile to tweak a small number of options and
get the least surprising behavior.