Use field number instead of memory position as `StorageKey` offset
This changes the method to compute the offset value of the `StorageKey`
for a particular storage access from using the memory offset of a given
field, which would result in collisions, to counting the number of
fields that precede it.

The reason for this change is that the memory offset calculation could
have multiple fields collide if they are zero-sized, as `StorageMap`s.

This is a breaking change because it will generate significantly
different IR for the same code.