Introduce unique `field_id` in `StorageKey`
This adds a field to `StorageKey` that is unique among storage fields.
It aims to alleviate the problem created by having multiple zero sized
fields inside of a storage struct, which are, by necessity, living in
the same slot.

It allows `StorageMap` (which itself is zero-sized) to address its own
values on a unique address map instead of colliding with adjacent
`StorageMap`s.

This is a breaking change because it will generate different IR
for the same code.

Fixes #4524
Fixes #4523