feat: improve newline insertion when comments are in between
The solution we adopted as a start would have us take the space between
leaf spans, except a problem was that a doc comment would be a leaf span
but a regular comment would not be a leaf span. Usually this would not
be an issue, but in certain cases where we perhaps want to comment code
out temporarily to test changes, the formatter would move comments around
unintentionally. For example, a doc comment followed by a commented out
function would cause the commented out function to move away from the doc
comment. Situations like these could ruin developer experience.

The main problem is that the entire whitespace gap between 2 leaf spans
could have multiple comments with multiple newline sequences, and all these
would be squished to the end of the first leaf span when inserting the
newlines.

The fix is to take 'snippets' while we're scanning for newline sequences
to insert and try to detect comments within these snippets. We then split
insertion of newline sequences into two phases: one before the first comment
(which is described above, which caused the issue above), one after
the last comment.