Implements where clauses for Structs and Enums.
This commit enables where clauses for Structs and Enums and also adds
two kinds of error handling.

One for when a type does not implement a trait and is used in a data structures
that requires that trait.

And another for when a generic type parameter should have the trait constraint.

Lets use the example below for demonstrating the two kinds of error handling.
```
enum MyEnum<T> where T: MyAdd {
  X: T,
}
```
The enum above has a constraint where `T` should implement the trait `MyAdd`.

When the type `MyEnum<u64>` is used and if `u64` does not implement the
trait `MyAdd` then the error `CompileError::TraitConstraintNotSatisfied`
is thrown.

The second type of error handling performed results in an error for the example below.
```
fn add2<G>(e: MyEnum<G>) -> G {
}
```
When `MyEnum<G>` is used and the type parameter `G` does not have the
trait constraint `where G: MyAdd` then the error `CompileError::TraitConstraintMissing`
is thrown.

Fixes #1159