Enables where clauses for traits.
With these changes where clauses are enabled for traits.
This commit also adds error handling when an `impl` is created for a generic
trait with where clause.

Generic traits are not supported yet on super traits or where clauses,
when the support is added we should also perform the bounds check there.

Fixes #4560
Fixes #970