ci-fix: use latest (sourced via path) version of std while testing sway reference examples (#4597)
## Description

closes #4595.
unblocks #4525.

Once #4596 is fixed we can remove std declarations in each manifest file
with a single patch statement in the workspace level manifest file. That
will also open the way to re-enable the disabled member (reentrancy) as
the example depends on sway-lib and sway-lib depends on latest released
std. So without the patch table it is not useable.

follow-up: #4596.