Fixes sha256 returning erroneous results.
The bug fixed was caused by sha256 being called with a byte size greater
than the string provided. This was caused by the intrinsic `__size_of`
returning the size padded to the next word alignment.

This was fixed by introducing a new intrinsic `__size_without_padding_of`
which does not return a padded size for strings.

Fixes #4491