Fixes module path issue in core.
Fixes issue where compiler would try to use module path `primitive_conversions::core::ops`.
With this fix the module path will be `ops`.
We also ignore `core` because it is in the name of the root module.

Test case is `core` specific and should be added to the repo once #4630 is merged.

Fixes issue found in #4630