make dir watcher use std::thread instead of async (#4658)
## Description
This PR changes how we are watching for directory changes in the
language server. Previously we were using tokio, this has now been
changed to use a regular thread and `std::mpsc::channel`. We are about
to move away from an `async` framework in the language server so this is
a small effort towards removing reliance on tokio.