Blanket trait implementations
This change allows blanket implementations for traits, i.e.:
```sway
impl<T> SomeTrait for T {}
```
As part of this we now import trait implementations that are in the same
namespace as an imported trait.

Ideally we'd want to import all existing trait implementations this way,
but this would require another pass or a query, which is probably out of
scope for a first implementation.

Traits are deduplicated here using the trait decl span.

Fix #4010