feat: make `--random-salt` default to forc-deploy and remove it as a flag (#4760)
## Description
closes #4758.

This PR makes --random-salt the default behaviour of forc-deploy, if
salt information is passed, the passed salt would be used but otherwise,
random-salt would be implied by default. Also adds --default-salt for CI
so that in CI we can always create the same contract id if we want to.
Useful for reproducible deployments.