fix: remove dependency on std for `regalloc_spill` unit test (#4775)
## Description
closes #4774.

This PR removes std library dependency for regalloc unit test so that we
can publish a new version without blocking our CI. In general we should
either not depend on std or we should only depend it by explicitly
depending on it by path. Otherwise we cannot bump version in our repo.