Fixes generic trait methods not found.
The issue this commit solves is the method not found error thrown when
we try to use generic traits.

This issue was fixed by calling `insert_trait_implementation_for_type` while resolving
tuples and arrays, we also call `insert_trait_implementation_for_type` as
a fail safe when a method is not found.

Fixes #4806
Unblocks #4701