Use relatives jumps (#4804)
## Description

Switch the compiler to use relative jumps instead of absolute jumps.

1. The function `relocate_control_flow` is now removed. It used to move
basic blocks around in case absolute addresses crossed the easily
addressable boundary. The code movement can then avoid having to load
the address to a register and using indirect jumps. This is unlikely to
happen with relative jumps. But if it does, the existing algorithm
wouldn't work anyway. So we can add a new one when required later.
2. In the case of function call jumps, we use `$pc` to save the return
address for the callee. This is a bit inefficient as it involves 3 extra
computations: `$$reta = ($pc - $is) / 4 + offset_to_return_point`. We
could change this to the following: Since we already have a common
return point for every function, compute the relative offset from that
point to the callee return point statically and use that. We don't have
the infrastructure to do that right now, but it shouldn't be too
difficult an extension.

Fixes #4443