chore: add forc-pkg codeowner (#4808)
## Description

We will be fixing #4709, but the one thing more important than fixing it
might be ensuring that it does not happen again. Until we have the
bandwidth to introduce nice CLI testing suite I am adding myself as code
owner to forc-pkg. Since forc-pkg lives between the compiler and forc,
it is seeing lots of traffic and I will try to make sure nothing changes
in terms of user experience during those changes. Feel free to close
this, if you feel like we might have a better way of ensuring this
before we got the CLI testing suite ready.