More error handling refactoring
This changes addresses the issues introduced in #4878

The `error_emitted` pattern is replaced by a scoping method on `Handler`
that collects errors within a closure and will error out if new erros are
introduced.

All instances of introducing `ErrorEmitted` as a literal value have
now been replaced by passing arround the receipt of the error emission
through the error representing values. Some types from the AST had to
be moved to `sway-types` to avoid circular dependencies introduced by
this change.

This also repairs some of the discrepancies to emitted warnings
introduced in #4878, as well as rename some of the `Handler` methods.

Fixes #4885
Fixes #4886