Refactor `parse_project` to return a result. (#4922)
## Description
This PR refactors the `parse_project` method to return a result instead
of writing to internal types in place. This helps to seperate the
parsing and writing logic into seperate phases, useful for future work
where we execute parsing on seperate threads.

I've also tried to clean up the `Session` interface at the same time.