Add benchmark framework to `sway-lsp` (#4927)
## Description
This PR adds a way to benchmark the LSP requests and `TokenMap` methods
using `criterion`. I've also added a large sway benchmark project that
can be used when running benchmarks now and in the future to measure
performance improvements.

Here is the result of the benchmarks. 

| Request Type         | duration  |
|----------------------|---------------|
| hover                | 442ms    |
documentHighlight    | 441ms   |
| definition           | 433ms   |
 prepareRename        | 433ms   |
| rename               | 436ms    |

| TokenMap method         | duration  |
|----------------------|---------------|
| tokens_for_file                | 230ms    |
idents_at_position    | 438ms   |
| tokens_at_position           | 437ms   |
 token_at_position        | 439ms   |
| parent_decl_at_position               | 437ms    |

related to #4901